Bazel: update `rules_python` and rework workspace loading
This updates `rules_python`. As `pip_install` was deprecated, this also
uses the new `pip_parse` in place of it, and sets up `requirements.txt`
as a locked version of `requirements.in`. Also, we now fix the python
version used by the build to 3.11.

Unfortunately, the new `rules_python` requires loading in 4 steps
(that is, a repository is defined, and then a symbol is `load`ed from
this repository, in turn defining another repository, and this 4 times).
Without using bzlmod, this means we need 4 different loading macros to
be able to load this into our internal workspace, up from the 2 we were
using (`codeql_workspace` and `codeql_workspace_deps`).

It _is_ messy. The only foreseeable clean solution to this is moving to
bzlmod that avoids this kind of gymnastics.

`codeql_workspace` and `codeql_workspace_deps` are left even if unused
for compatibility with the current internal workspace.