Ruby: Model ERB flow for ActionController
Extend the existing modeling for ViewComponent to ActionController.

The `render` method in ActionController behaves in a similar way except
that the ERB file is evaluated in the context of the controller class,
rather than a separate view class. This means we need the value of
`self` at the `render` call to flow to the `self` in the ERB file.